Fix pom project coordinates to follow the right order of naming
