+++ src/main/java/com/horus/yoga/HorusApplication.java
@@ -1,0 +1,13 @@
+package com.horus.yoga;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HorusApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HorusApplication.class, args);
+	}
+
+}
