+++ src/main/java/com/horus/yoga/constant/Constant.java
@@ -1,0 +1,10 @@
+package com.horus.yoga.constant;
+
+public class Constant {
+
+    public static <T> String Deleted(T item){
+        return String.format("%s deleted", item);
+    }
+
+    public static final String SUCCESS = "Success";
+}
