+++ src/main/java/com/horus/yoga/controller/BaseController.java
@@ -1,0 +1,17 @@
+package com.horus.yoga.controller;
+
+import com.horus.yoga.util.APIResponse;
+import java.util.List;
+import java.util.UUID;
+
+public interface BaseController<Entity, DTO> {
+    APIResponse<DTO> getById(UUID id);
+
+    APIResponse<List<DTO>> getAll();
+
+    APIResponse<DTO> create(Entity object);
+
+    APIResponse<DTO> update(UUID id, Entity object);
+
+    APIResponse<String> delete(UUID id);
+}
