+++ src/main/java/com/horus/yoga/dtoMapper/UserMapper.java
@@ -1,0 +1,14 @@
+package com.horus.yoga.dtoMapper;
+
+import com.horus.yoga.dto.UserDTO;
+import com.horus.yoga.entity.User;
+import org.mapstruct.Mapper;
+
+
+
+@Mapper (componentModel = "spring")
+public interface UserMapper {
+    UserDTO userToUserDTO(User user);
+    User userDTOToUser(UserDTO userDTO);
+
+}
