+++ src/main/java/com/horus/yoga/entity/Payment.java
@@ -1,0 +1,73 @@
+package com.horus.yoga.entity;
+
+import com.horus.yoga.config.Auditable;
+import com.horus.yoga.enums.PaymentMethod;
+import jakarta.persistence.*;
+
+import java.util.Date;
+import java.util.UUID;
+
+@Entity
+@Table(name = "payment")
+public class Payment extends Auditable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+
+
+    @ManyToOne
+    @JoinColumn(name = "user_id",
+                referencedColumnName = "id")
+    private User userId;
+
+
+    @ManyToOne
+    @JoinColumn(name = "package_id",
+                referencedColumnName = "id")
+    private Bundle packageId;
+
+    @Column(nullable = false)
+    private int amount;
+
+    @Column(name ="payment_date", nullable = false)
+    private Date paymentDate;
+
+    @Column(name = "payment_method",nullable = false)
+    @Enumerated(EnumType.STRING)
+    private PaymentMethod paymentMethod;
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(UUID id) {
+        this.id = id;
+    }
+
+    public int getAmount() {
+        return amount;
+    }
+
+    public void setAmount(int amount) {
+        this.amount = amount;
+    }
+
+    public Date getPaymentDate() {
+        return paymentDate;
+    }
+
+    public void setPaymentDate(Date paymentDate) {
+        this.paymentDate = paymentDate;
+    }
+
+    public PaymentMethod getPaymentMethod() {
+        return paymentMethod;
+    }
+
+    public void setPaymentMethod(PaymentMethod paymentMethod) {
+        this.paymentMethod = paymentMethod;
+    }
+
+
+}
