+++ src/main/java/com/horus/yoga/enums/PaymentMethod.java
@@ -1,0 +1,11 @@
+package com.horus.yoga.enums;
+
+public enum PaymentMethod {
+    CASH,
+    VISA,
+    PAYPAL,
+    MASTERCARD,
+    STRIPE,
+    BITCOIN,
+    WIRETRANSFER;
+}
