+++ src/main/java/com/horus/yoga/enums/Role.java
@@ -1,0 +1,6 @@
+package com.horus.yoga.enums;
+
+public enum Role {
+    STUDENT,
+    INSTRUCTOR;
+}
