+++ src/main/java/com/horus/yoga/repository/AttendanceRepository.java
@@ -1,0 +1,9 @@
+package com.horus.yoga.repository;
+
+import com.horus.yoga.entity.Attendance;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface AttendanceRepository extends JpaRepository<Attendance, UUID> {
+}
