+++ src/main/java/com/horus/yoga/repository/BundleRepository.java
@@ -1,0 +1,9 @@
+package com.horus.yoga.repository;
+
+import com.horus.yoga.entity.Bundle;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface BundleRepository extends JpaRepository<Bundle,UUID> {
+}
