+++ src/main/java/com/horus/yoga/repository/PaymentRepository.java
@@ -1,0 +1,9 @@
+package com.horus.yoga.repository;
+
+import com.horus.yoga.entity.Payment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface PaymentRepository extends JpaRepository<Payment, UUID> {
+}
