+++ src/main/java/com/horus/yoga/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.horus.yoga.repository;
+
+import com.horus.yoga.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+}
