+++ src/main/java/com/horus/yoga/util/AttendanceId.java
@@ -1,0 +1,32 @@
+package com.horus.yoga.util;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+
+import java.util.UUID;
+
+@Embeddable
+public class AttendanceId {
+
+    @Column(name = "user_id")
+    private UUID userId;
+
+    @Column(name = "class_id")
+    private UUID yogaClassId;
+
+    public UUID getUserId() {
+        return userId;
+    }
+
+    public void setUserId(UUID userId) {
+        this.userId = userId;
+    }
+
+    public UUID getYogaClassId() {
+        return yogaClassId;
+    }
+
+    public void setYogaClassId(UUID yogaClassId) {
+        this.yogaClassId = yogaClassId;
+    }
+}
