--- src/main/java/com/yoga/horus/HorusApplication.java
@@ -1,13 +1,0 @@
-package com.yoga.horus;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HorusApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(HorusApplication.class, args);
-	}
-
-}
