--- src/main/java/com/yoga/horus/config/AuditConfig.java
@@ -1,17 +1,0 @@
-package com.yoga.horus.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing(auditorAwareRef = "auditorAware")
-public class AuditConfig {
-
-    @Bean
-    public AuditorAware<String> auditorAware() {
-        return new AuditorAwareImpl();
-    }
-
-}
