--- src/main/java/com/yoga/horus/constant/Constant.java
@@ -1,10 +1,0 @@
-package com.yoga.horus.constant;
-
-public class Constant {
-
-    public static <T> String Deleted(T item){
-        return String.format("%s deleted", item);
-    }
-
-    public static final String SUCCESS = "Success";
-}
