--- src/main/java/com/yoga/horus/controller/BaseController.java
@@ -1,17 +1,0 @@
-package com.yoga.horus.controller;
-
-import com.yoga.horus.util.APIResponse;
-import java.util.List;
-import java.util.UUID;
-
-public interface BaseController<Entity, DTO> {
-    APIResponse<DTO> getById(UUID id);
-
-    APIResponse<List<DTO>> getAll();
-
-    APIResponse<DTO> create(Entity object);
-
-    APIResponse<DTO> update(UUID id, Entity object);
-
-    APIResponse<String> delete(UUID id);
-}
