--- src/main/java/com/yoga/horus/dto/UserDTO.java
@@ -1,12 +1,0 @@
-package com.yoga.horus.dto;
-
-import com.yoga.horus.enums.Role;
-
-import java.util.UUID;
-
-
-public record UserDTO(UUID id,
-                      String firstName,
-                      String email,
-                      Role role) {
-}
