--- src/main/java/com/yoga/horus/enums/PaymentMethod.java
@@ -1,11 +1,0 @@
-package com.yoga.horus.enums;
-
-public enum PaymentMethod {
-    CASH,
-    VISA,
-    PAYPAL,
-    MASTERCARD,
-    STRIPE,
-    BITCOIN,
-    WIRETRANSFER;
-}
