--- src/main/java/com/yoga/horus/enums/Role.java
@@ -1,6 +1,0 @@
-package com.yoga.horus.enums;
-
-public enum Role {
-    STUDENT,
-    INSTRUCTOR;
-}
