--- src/main/java/com/yoga/horus/repository/AttendanceRepository.java
@@ -1,9 +1,0 @@
-package com.yoga.horus.repository;
-
-import com.yoga.horus.entity.Attendance;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface AttendanceRepository extends JpaRepository<Attendance, UUID> {
-}
