--- src/main/java/com/yoga/horus/repository/BundleRepository.java
@@ -1,9 +1,0 @@
-package com.yoga.horus.repository;
-
-import com.yoga.horus.entity.Bundle;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface BundleRepository extends JpaRepository<Bundle,UUID> {
-}
