--- src/main/java/com/yoga/horus/repository/PaymentRepository.java
@@ -1,9 +1,0 @@
-package com.yoga.horus.repository;
-
-import com.yoga.horus.entity.Payment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface PaymentRepository extends JpaRepository<Payment, UUID> {
-}
