--- src/main/java/com/yoga/horus/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.yoga.horus.repository;
-
-import com.yoga.horus.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface UserRepository extends JpaRepository<User, UUID> {
-}
