--- src/main/java/com/yoga/horus/repository/YogaClassRepository.java
@@ -1,11 +1,0 @@
-package com.yoga.horus.repository;
-
-import com.yoga.horus.entity.YogaClass;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-
-public interface YogaClassRepository extends JpaRepository<YogaClass, UUID> {
-
-}
